Adds user controller
